First pass at conversion BPM int to double. Not working, Implement Sequencer64 type precision setting and import... Thanks to Chris Ahlstrom.

+++ src/midibus.h
@@ -199,7 +199,7 @@ private:
     int m_queue;
 
     int m_ppqn;
-    int m_bpm;
+    double m_bpm;
 
     int  m_num_poll_descriptors;
     struct pollfd *m_poll_descriptors;
@@ -241,9 +241,9 @@ public:
     int get_num_out_buses();
     int get_num_in_buses();
 
-    void set_bpm(int a_bpm);
+    void set_bpm(double a_bpm);
     void set_ppqn(int a_ppqn);
-    int get_bpm()
+    double get_bpm()
     {
         return m_bpm;
     }
